requirements.txt
black
mypy
scikit-learn
Pillow
numpy
rich
